full functionality finished
full CRUD functionality is implemented, also upon Creating a new comment, replying, deleting, updating everything will be saved to local storage to be fetched in the next reload. At this point the animations are half implemented, the animations are fully functional at the main comments level but still not applied on replies, this would be the next step with some arrangment to the code.